# Bandwidth-preserving branches in row/col broadcast