This helps with performance in `A .* b` type cases

With arrays
```julia
A = BandedMatrix(0=>Float64[1:3000;]); v = rand(size(A,1));
```
| Operation | Master | PR |
| --- | :-: | :-: |
| `A .* v` | 50.560 μs | 19.609 μs |
| `v .* A` | 49.867 μs | 23.308 μs |
| `A .* v'` | 51.944 μs| 10.112 μs |
| `v' .* A` | 58.198 μs| 10.388 μs |